Pull out MetadataStore into a separate file.
