+++ src/api_server.cc
@@ -36,86 +36,6 @@ namespace http = boost::network::http;
 
 namespace google {
 
-MetadataStore::Metadata MetadataStore::Metadata::IGNORED() {
-  return MetadataStore::Metadata();
-}
-
-MetadataStore::MetadataStore(const MetadataAgentConfiguration& config)
-    : config_(config) {}
-
-void MetadataStore::UpdateResource(const std::vector<std::string>& resource_ids,
-                                   const MonitoredResource& resource) {
-  std::lock_guard<std::mutex> lock(resource_mu_);
-  // TODO: How do we handle deleted resources?
-  // TODO: Do we care if the value was already there?
-  for (const std::string& id : resource_ids) {
-    if (config_.VerboseLogging()) {
-      LOG(INFO) << "Updating resource map '" << id << "'->" << resource;
-    }
-    resource_map_.emplace(id, resource);
-  }
-}
-
-void MetadataStore::UpdateMetadata(const MonitoredResource& resource,
-                                   Metadata&& entry) {
-  std::lock_guard<std::mutex> lock(metadata_mu_);
-  if (config_.VerboseLogging()) {
-    LOG(INFO) << "Updating metadata map " << resource << "->{"
-              << "version: " << entry.version << ", "
-              << "is_deleted: " << entry.is_deleted << ", "
-              << "created_at: " << time::rfc3339::ToString(entry.created_at) << ", "
-              << "collected_at: " << time::rfc3339::ToString(entry.collected_at)
-              << ", "
-              << "metadata: " << *entry.metadata << ", "
-              << "ignore: " << entry.ignore
-              << "}";
-  }
-  // Force value update. The repeated search is inefficient, but shouldn't
-  // be a huge deal.
-  metadata_map_.erase(resource);
-  metadata_map_.emplace(resource, std::move(entry));
-}
-
-std::map<MonitoredResource, MetadataStore::Metadata>
-    MetadataStore::GetMetadataMap() const {
-  std::lock_guard<std::mutex> lock(metadata_mu_);
-
-  std::map<MonitoredResource, Metadata> result;
-  for (const auto& kv : metadata_map_) {
-    const MonitoredResource& resource = kv.first;
-    const Metadata& metadata = kv.second;
-    result.emplace(resource, metadata.Clone());
-  }
-  return result;
-}
-
-void MetadataStore::PurgeDeletedEntries() {
-  std::lock_guard<std::mutex> lock(metadata_mu_);
-
-  for (auto it = metadata_map_.begin(); it != metadata_map_.end(); ) {
-    const MonitoredResource& resource = it->first;
-    const Metadata& entry = it->second;
-    if (entry.is_deleted) {
-      if (config_.VerboseLogging()) {
-        LOG(INFO) << "Purging metadata entry " << resource << "->{"
-                  << "version: " << entry.version << ", "
-                  << "is_deleted: " << entry.is_deleted << ", "
-                  << "created_at: " << time::rfc3339::ToString(entry.created_at)
-                  << ", "
-                  << "collected_at: " << time::rfc3339::ToString(entry.collected_at)
-                  << ", "
-                  << "metadata: " << *entry.metadata << ", "
-                  << "ignore: " << entry.ignore
-                  << "}";
-      }
-      it = metadata_map_.erase(it);
-    } else {
-      ++it;
-    }
-  }
-}
-
-
 class MetadataApiServer {
  public:
   MetadataApiServer(const MetadataAgent& agent, int server_threads,
